requirements.txt
pyTelegramBotAPI~=4.26.0
python-dotenv~=1.0.1
openpyxl~=3.1.5
mysql-connector-python~=9.2.0
Flask~=3.1.0
Flask-JWT-Extended~=4.7.1
imgurpython~=1.1.7